fix: pm positions
