+++ packages/layout-engine/pm-adapter/src/index.ts
@@ -40,8 +40,8 @@ export { SectionType } from './types.js';
 // Re-export public API functions from internal implementation
 export { toFlowBlocks, toFlowBlocksMap } from './internal.js';
 
-// Re-export run type guards
-export { isTextRun } from './converters/paragraph.js';
+// Re-export run type guards and run utilities
+export { isTextRun, expandRunsForInlineNewlines } from './converters/paragraph.js';
 
 // Re-export cache for incremental conversion
 export { FlowBlockCache } from './cache.js';
